+++ pyproject.toml
@@ -7,7 +7,7 @@ python-tag = "py310"  # this fixes the python tag in the built wheel to py310 in
 
 [project]
 version = "0.0.1"
-name = "tempo-worklog-creator"
+name = "tempo-worklog-cli"
 description = "CLI for creating tempo worklogs on Jira"
 readme = "README.md"
 requires-python = "==3.10.*"  # pin to 3.10
@@ -23,7 +23,7 @@ dependencies = [
     "tempo-api-python-client",
 ]
 [project.scripts]
-tempo = "tempo_worklog_creator.cli:cli"
+tempo = "tempo_worklog_cli.cli:cli"
 
 [project.optional-dependencies]
 dev = [
