+++ src/tempo_worklog_cli/cli.py
@@ -6,9 +6,9 @@ import click
 from click import Context
 from dotenv import load_dotenv
 
-from tempo_worklog_creator.time_span import TimeSpan
-from tempo_worklog_creator.util.serialization import converter
-from tempo_worklog_creator.worklog_creator import WorkLogCreator
+from tempo_worklog_cli.time_span import TimeSpan
+from tempo_worklog_cli.util.serialization import converter
+from tempo_worklog_cli.worklog_creator import WorkLogCreator
 
 load_dotenv()
